feat: create folder structure
