+++ app/src/main/java/dog/snow/androidrecruittest/data/model/Detail.kt
@@ -1,4 +1,4 @@
-package dog.snow.androidrecruittest.ui.model
+package dog.snow.androidrecruittest.data.model
 
 import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
