+++ app/src/main/java/dog/snow/androidrecruittest/data/model/RawPhoto.kt
@@ -1,12 +1,13 @@
-package dog.snow.androidrecruittest.ui.model
+package dog.snow.androidrecruittest.data.model
 
 import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
 
 @Parcelize
-data class ListItem(
+data class RawPhoto(
     val id: Int,
+    val albumId: Int,
     val title: String,
-    val albumTitle: String,
+    val url: String,
     val thumbnailUrl: String
 ) : Parcelable
