+++ app/src/main/java/dog/snow/androidrecruittest/data/model/RawUser.kt
@@ -1,4 +1,4 @@
-package dog.snow.androidrecruittest.repository.model
+package dog.snow.androidrecruittest.data.model
 
 import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
