+++ app/src/main/java/dog/snow/androidrecruittest/data/source/remote/service/AlbumService.kt
@@ -1,0 +1,5 @@
+package dog.snow.androidrecruittest.data.source.remote.service
+
+interface AlbumService {
+
+}
