--- app/src/main/java/dog/snow/androidrecruittest/repository/service/PhotoService.kt
@@ -1,7 +1,0 @@
-package dog.snow.androidrecruittest.repository.service
-
-import dog.snow.androidrecruittest.repository.model.RawPhoto
-
-interface PhotoService {
-
-}
