--- app/src/main/java/dog/snow/androidrecruittest/repository/service/UserService.kt
@@ -1,7 +1,0 @@
-package dog.snow.androidrecruittest.repository.service
-
-import dog.snow.androidrecruittest.repository.model.RawUser
-
-interface UserService {
-
-}
