+++ app/src/main/java/dog/snow/androidrecruittest/ui/details/DetailsFragment.kt
@@ -1,4 +1,4 @@
-package dog.snow.androidrecruittest.ui
+package dog.snow.androidrecruittest.ui.details
 
 import androidx.fragment.app.Fragment
 import dog.snow.androidrecruittest.R
