+++ app/src/main/java/dog/snow/androidrecruittest/ui/list/ListFragment.kt
@@ -1,4 +1,4 @@
-package dog.snow.androidrecruittest.ui
+package dog.snow.androidrecruittest.ui.list
 
 import androidx.fragment.app.Fragment
 import dog.snow.androidrecruittest.R
