+++ app/src/main/java/dog/snow/androidrecruittest/ui/main/MainActivity.kt
@@ -1,7 +1,8 @@
-package dog.snow.androidrecruittest
+package dog.snow.androidrecruittest.ui.main
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import dog.snow.androidrecruittest.R
 
 class MainActivity : AppCompatActivity(R.layout.main_activity){
     override fun onCreate(savedInstanceState: Bundle?) {
